updated RPC import
