+++ sandbox/src/app/constants/rpc.ts
@@ -1,0 +1,6 @@
+export const RPC_URLS: Record<'devnet' | 'mainnet' | 'sonic-mainnet' | 'sonic-testnet', string> = {
+  'devnet': 'https://api.devnet.solana.com',
+  'mainnet': 'https://api.mainnet-beta.solana.com',
+  'sonic-mainnet': 'https://mainnet.rpc.sonic.so',
+  'sonic-testnet': 'https://testnet.rpc.sonic.so'
+};
